src/api/requirements-testing.txt
# Florgon API server requirements.

# Web framework (FastAPI).
fastapi==0.88.0
fastapi-cache2[redis]==0.1.9
fastapi_mail==1.2.4

# ASGI server worker with
# server process manager.
uvicorn==0.20.0
gunicorn==20.1.0
httpx==0.23.3

# Florgon Gatey SDK.
# (Error logging service).
gatey-sdk==0.0.7

# ORM with Database engine.
sqlalchemy==1.4.45  # ORM.
psycopg2-binary==2.9.5  # SQL Engine (PostgreSQL engine core).
aioredis==2.0.1  # Async Redis engine.

# Some utils libraries.
validate_email==1.3  # Used for sign-up process.

# Security, authentication.
pyjwt==2.5.0  # JWTokens implementation for the base token.
pyotp==2.8.0  # OTP for 2FA. Used for email and mobile 2FA process.

# Testing and OAuth providers (make request to OAuth provider OAuth code resolver).
requests==2.28.1  # Make requests to server.

# CI/CD, Testing.
pytest==7.2.0  # Unit tests.
pytest-cov==4.0.0  # Tests coverage.
